added loading

+++ src/components/InventoryItems.jsx
@@ -1,11 +1,15 @@
-import { useEffect, useContext } from "react";
+import { useEffect, useContext, useState } from "react";
 import ItemsContext from "../features/ItemsContext";
 import itemService from "../services/itemService";
 import PriceContext from "../features/PriceContext";
+import LoadingContext from "../features/LoadingContext";
+import LoadingSpinner from "../loadingComponents/LoadingSpinner";
 
 function InventoryItems() {
   const { items, setItems } = useContext(ItemsContext);
-  const { price, setPrice } = useContext(PriceContext);
+  const { setPrice } = useContext(PriceContext);
+  const { loading, setLoading } = useContext(LoadingContext);
+
   // // let newNumber = 0;
 
   useEffect(() => {
@@ -16,12 +20,14 @@ function InventoryItems() {
   }, []);
 
   const deleteItem = (id) => {
+    setLoading(true);
     itemService
       .deleteItem(id)
       .then((_res) => {
         setItems(items.filter((item) => item.id !== id));
       })
-      .catch((error) => console.log(error));
+      .catch((error) => console.log(error))
+      .finally(() => setLoading(false));
   };
 
   const handleDecreaseQuantity = (id, item) => {
@@ -52,37 +58,55 @@ function InventoryItems() {
     setPrice((prevState) => [...prevState, priceToAdd]);
   };
 
+  if (loading) {
+    return (
+      <div className="flex flex-col  items-center justify-center">
+        <LoadingSpinner />
+      </div>
+    );
+  }
+
   return (
     <div className="bg-darkgreen grid ">
       <ul className="grid grid-cols-6 gap-3 p-2">
-        {items.map((item) => (
-          <li
-            key={item.id}
-            className="flex-column p-3 bg-primary gap-3 rounded-lg"
-          >
-            <img src={item.photoInfo.url} alt="" className="w-52" />
-            <p className="text-lg">{item.name}</p>
-            <p>Quantity: {item.quantity}</p>
-            <p>₱{item.price}</p>
-            <div className="flex gap-2 justify-evenly">
-              <button
-                className="bg-secondary w-1/2 rounded-md cursor-pointer hover:bg-green focus:outline-none sell"
-                onClick={() => {
-                  handleDecreaseQuantity(item.id, item);
-                  handleGetPrice(item);
-                }}
-              >
-                Sell
-              </button>
-              <button
-                className="bg-red  w-1/2 rounded-md cursor-pointer hover:bg-orange-500 remove"
-                onClick={() => deleteItem(item.id)}
-              >
-                Remove
-              </button>
-            </div>
-          </li>
-        ))}
+        {items.length === 0 ? (
+          <div className="flex flex-col items-center justify-center text-center w-max">
+            <h1 className="page-title-txt">
+              MY<span>ASENSO</span>
+            </h1>
+
+            <p>Add your first item by clicking the "add item" button</p>
+          </div>
+        ) : (
+          items.map((item) => (
+            <li
+              key={item.id}
+              className="flex-column p-3 bg-primary gap-3 rounded-lg"
+            >
+              <img src={item.photoInfo.url} alt="" className="w-52" />
+              <p className="text-lg">{item.name}</p>
+              <p>Quantity: {item.quantity}</p>
+              <p>₱{item.price}</p>
+              <div className="flex gap-2 justify-evenly">
+                <button
+                  className="bg-secondary w-1/2 rounded-md cursor-pointer hover:bg-green focus:outline-none sell"
+                  onClick={() => {
+                    handleDecreaseQuantity(item.id, item);
+                    handleGetPrice(item);
+                  }}
+                >
+                  Sell
+                </button>
+                <button
+                  className="bg-red  w-1/2 rounded-md cursor-pointer hover:bg-orange-500 remove"
+                  onClick={() => deleteItem(item.id)}
+                >
+                  Remove
+                </button>
+              </div>
+            </li>
+          ))
+        )}
       </ul>
     </div>
   );
